app/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
kafka-python
requests
